// src/types/enums/dialog.enum.ts
export enum DialogEnum {
  INIT = "INIT",
  DIALOG_USER = "DIALOG_USER",
  DIALOG_RESORUCE = "DIALOG_RESORUCE",
  DIALOG_MODULE_SYSTEM = "DIALOG_MODULE_SYSTEM",
  DIALOG_ROLE = "DIALOG_ROLE",
  DIALOG_TIPODOC = "DIALOG_TIPODOC",
  DIALOG_EMPRESA = "DIALOG_EMPRESA",
  DIALOG_SERIES = "DIALOG_SERIES",
  DIALOG_SERIES_MIGRATE = "DIALOG_SERIES_MIGRATE",
  OPEN_EDIT = "OPEN_EDIT",
}